+++ app/views/hyrax/admin/workflows/index.html.erb
@@ -4,90 +4,15 @@
 
 <div class="row">
   <div class="col-md-12">
-    <div class="panel panel-default tabs">
-      <ul class="nav nav-tabs" role="tablist">
-        <li class="active">
-          <a href="#under-review" role="tab" data-toggle="tab"><%= t('.tabs.under_review') %></a>
-        </li>
-        <li>
-          <a href="#published" role="tab" data-toggle="tab"><%= t('.tabs.published') %></a>
-        </li>
-      </ul>
-      <div class="tab-content">
-        <div id="under-review" class="tab-pane active">
-          <div class="panel panel-default labels">
-            <div class="panel-body">
-              <div class="table-responsive">
-                <table class="table table-condensed table-striped datatable">
-                  <thead>
-                    <tr>
-                      <th width="40%">Work</th>
-                      <th width="20%">Depositor</th>
-                      <th width="20%">Submission Date</th>
-                      <th width="20%">Status</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    <% @status_list.each do |document| %>
-                      <% next unless can? :review, document %>
-                      <tr>
-                        <td>
-                          <%= link_to document, [main_app, document] %>
-                        </td>
-                        <td>
-                          <%= safe_join(document.creator, tag(:br)) %>
-                        </td>
-                        <td>
-                          <%= document.date_modified %>
-                        </td>
-                        <td>
-                          <span class="state state-pending"><%= document.workflow_state %></span>
-                        </td>
-                      </tr>
-                    <% end %>
-                  </tbody>
-                </table>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div id="published" class="tab-pane">
-          <div class="panel panel-default labels">
-            <div class="panel-body">
-              <div class="table-responsive">
-                <table class="table table-condensed table-striped datatable">
-                  <thead>
-                    <tr>
-                      <th width="40%">Work</th>
-                      <th width="20%">Depositor</th>
-                      <th width="20%">Submission Date</th>
-                      <th width="20%">Status</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    <% @published_list.each do |document| %>
-                      <% next unless can? :review, document %>
-                      <tr>
-                        <td>
-                          <%= link_to document, [main_app, document] %>
-                        </td>
-                        <td>
-                          <%= safe_join(document.creator, tag(:br)) %>
-                        </td>
-                        <td>
-                          <%= document.date_modified %>
-                        </td>
-                        <td>
-                          <span class="state state-pending"><%= document.workflow_state %></span>
-                        </td>
-                      </tr>
-                    <% end %>
-                  </tbody>
-                </table>
-              </div>
-            </div>
-          </div>
-        </div>
+    <div id="content" class="col-md-9 col-sm-8">
+      <%= render 'constraints' %>
+
+      <div id="sidebar" class="col-md-3 col-sm-4">
+        <%= render 'search_sidebar' %>
+      </div>
+
+      <div id="content" class="col-md-9 col-sm-8">
+        <%= render 'search_results' %>
       </div>
     </div>
   </div>
